+++ src/blocks/__TEMPLATE__/index.tsx
@@ -1,8 +1,8 @@
 import type { InterpretedAttributes } from "./attributes";
 import type { InterpretedUsedContext } from "./context";
 import { registerBlockType } from "@atomicsmash/blocks-helpers";
-import { registerLaunchpadBlocksCollection } from "@plugin/blocks/helpers.editor";
 import { registerBlockVariation } from "@wordpress/blocks";
+import { registerLaunchpadBlocksCollection } from "@plugin/blocks/helpers.editor";
 // import { Icon } from "@blocks/svgs";
 import blockMetaData from "./block.json";
 import { deprecated } from "./deprecation";